+++ datafusion/expr/src/built_in_function.rs
@@ -119,10 +119,6 @@ pub enum BuiltinScalarFunction {
     ArrayReplaceN,
     /// array_replace_all
     ArrayReplaceAll,
-    /// array_intersect
-    ArrayIntersect,
-    /// array_union
-    ArrayUnion,
     /// array_except
     ArrayExcept,
 
@@ -287,8 +283,6 @@ impl BuiltinScalarFunction {
             BuiltinScalarFunction::ArrayReplace => Volatility::Immutable,
             BuiltinScalarFunction::ArrayReplaceN => Volatility::Immutable,
             BuiltinScalarFunction::ArrayReplaceAll => Volatility::Immutable,
-            BuiltinScalarFunction::ArrayIntersect => Volatility::Immutable,
-            BuiltinScalarFunction::ArrayUnion => Volatility::Immutable,
             BuiltinScalarFunction::Ascii => Volatility::Immutable,
             BuiltinScalarFunction::BitLength => Volatility::Immutable,
             BuiltinScalarFunction::Btrim => Volatility::Immutable,
@@ -356,24 +350,6 @@ impl BuiltinScalarFunction {
             BuiltinScalarFunction::ArrayReplace => Ok(input_expr_types[0].clone()),
             BuiltinScalarFunction::ArrayReplaceN => Ok(input_expr_types[0].clone()),
             BuiltinScalarFunction::ArrayReplaceAll => Ok(input_expr_types[0].clone()),
-            BuiltinScalarFunction::ArrayIntersect => {
-                match (input_expr_types[0].clone(), input_expr_types[1].clone()) {
-                    (DataType::Null, DataType::Null) | (DataType::Null, _) => {
-                        Ok(DataType::Null)
-                    }
-                    (_, DataType::Null) => {
-                        Ok(List(Arc::new(Field::new("item", Null, true))))
-                    }
-                    (dt, _) => Ok(dt),
-                }
-            }
-            BuiltinScalarFunction::ArrayUnion => {
-                match (input_expr_types[0].clone(), input_expr_types[1].clone()) {
-                    (DataType::Null, dt) => Ok(dt),
-                    (dt, DataType::Null) => Ok(dt),
-                    (dt, _) => Ok(dt),
-                }
-            }
             BuiltinScalarFunction::ArrayExcept => {
                 match (input_expr_types[0].clone(), input_expr_types[1].clone()) {
                     (DataType::Null, _) | (_, DataType::Null) => {
@@ -553,9 +529,6 @@ impl BuiltinScalarFunction {
             BuiltinScalarFunction::ArrayReplaceAll => {
                 Signature::any(3, self.volatility())
             }
-            BuiltinScalarFunction::ArrayIntersect => Signature::any(2, self.volatility()),
-            BuiltinScalarFunction::ArrayUnion => Signature::any(2, self.volatility()),
-
             BuiltinScalarFunction::Concat
             | BuiltinScalarFunction::ConcatWithSeparator => {
                 Signature::variadic(vec![Utf8], self.volatility())
@@ -876,10 +849,6 @@ impl BuiltinScalarFunction {
             }
             BuiltinScalarFunction::ArrayReplaceAll => {
                 &["array_replace_all", "list_replace_all"]
-            }
-            BuiltinScalarFunction::ArrayUnion => &["array_union", "list_union"],
-            BuiltinScalarFunction::ArrayIntersect => {
-                &["array_intersect", "list_intersect"]
             }
             BuiltinScalarFunction::OverLay => &["overlay"],
         }
