+++ datafusion/functions-array/src/lib.rs
@@ -30,10 +30,11 @@ pub mod macros;
 
 mod array_has;
 mod concat;
+mod core;
 mod extract;
 mod kernels;
-mod make_array;
 mod rewrite;
+mod set_ops;
 mod udf;
 mod utils;
 
@@ -51,13 +52,15 @@ pub mod expr_fn {
     pub use super::concat::array_append;
     pub use super::concat::array_concat;
     pub use super::concat::array_prepend;
+    pub use super::core::make_array;
     pub use super::extract::array_element;
     pub use super::extract::array_pop_back;
     pub use super::extract::array_pop_front;
     pub use super::extract::array_slice;
-    pub use super::make_array::make_array;
+    pub use super::set_ops::array_distinct;
+    pub use super::set_ops::array_intersect;
+    pub use super::set_ops::array_union;
     pub use super::udf::array_dims;
-    pub use super::udf::array_distinct;
     pub use super::udf::array_empty;
     pub use super::udf::array_length;
     pub use super::udf::array_ndims;
@@ -90,7 +93,7 @@ pub fn register_all(registry: &mut dyn FunctionRegistry) -> Result<()> {
         extract::array_pop_back_udf(),
         extract::array_pop_front_udf(),
         extract::array_slice_udf(),
-        make_array::make_array_udf(),
+        core::make_array_udf(),
         array_has::array_has_udf(),
         array_has::array_has_all_udf(),
         array_has::array_has_any_udf(),
@@ -98,10 +101,12 @@ pub fn register_all(registry: &mut dyn FunctionRegistry) -> Result<()> {
         udf::array_length_udf(),
         udf::flatten_udf(),
         udf::array_sort_udf(),
-        udf::array_distinct_udf(),
         udf::array_repeat_udf(),
         udf::array_resize_udf(),
         udf::array_reverse_udf(),
+        set_ops::array_distinct_udf(),
+        set_ops::array_intersect_udf(),
+        set_ops::array_union_udf(),
     ];
     functions.into_iter().try_for_each(|udf| {
         let existing_udf = registry.register_udf(udf)?;
